requirements.txt
dnspython==2.7.0
numpy==2.0.2
pandas==2.3.3
psycopg2-binary==2.9.11
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.3
